Fix broken stop functionality.
The stop behavior wasn't actually working, and
the tests were lying about it.  Because they
uses ivar state, the threads in the tests dup'd
the same queue.  However with truly separate
producers and workers, this wouldn't work.  Fixed
by passing QSTOP from producer to workers as stop
command, and duping the the queue in one of the
halt tests to avoid state problems.